+++ jest.config.js
@@ -25,9 +25,8 @@ module.exports = {
     '^.+\\.(ts|tsx)$': [
       'ts-jest',
       {
-        tsconfig: '<rootDir>/tsconfig.json',
-        diagnostics: false,
-        isolatedModules: true
+        tsconfig: '<rootDir>/tsconfig.jest.json',
+        diagnostics: false
       }
     ]
   },